Adding code from plab3 for connmgr (Req 3: #3)
This code is *not* ready to be used, it still needs a lot of changes in
order to work properly for the final project. For examle, it is not
properly  set up to be multithreaded and it can not currently access the
logger functions defined in main. It also needs access to an sbuffer
implementation.